refactor: move restore_positions, save_positions and get_ids into utils
